+++ src/bespokeasm/assembler/keywords.py
@@ -8,7 +8,7 @@ BYTECODE_DIRECTIVES_SET = {
 }
 
 PREPROCESSOR_DIRECTIVES_SET = {
-    'include', 'require', 'create_memzone', 'print',
+    'include', 'require', 'error', 'create_memzone', 'print',
     'define', 'if', 'elif', 'else', 'endif', 'ifdef', 'ifndef',
     'mute', 'unmute', 'emit',
     'create-scope', 'use-scope', 'deactivate-scope',