+++ test/test_configgen.py
@@ -388,7 +388,7 @@ class TestConfigurationGeneration(unittest.TestCase):
         self._assert_grouped_item_list(
             item_match_str,
             [
-                'include', 'require', 'create_memzone', 'print', 'define', 'if',
+                'include', 'require', 'error', 'create_memzone', 'print', 'define', 'if',
                 'elif', 'else', 'endif', 'ifdef', 'ifndef',
                 'mute', 'unmute', 'emit',
                 'create-scope', 'use-scope', 'deactivate-scope',
@@ -501,7 +501,7 @@ class TestConfigurationGeneration(unittest.TestCase):
         self._assert_grouped_item_list(
             item_match_str,
             [
-                'include', 'require', 'create_memzone', 'print', 'define',
+                'include', 'require', 'error', 'create_memzone', 'print', 'define',
                 'if', 'elif', 'else', 'endif', 'ifdef', 'ifndef',
                 'mute', 'unmute', 'emit',
                 'create-scope', 'use-scope', 'deactivate-scope',