+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Changes that are planned but not implemented yet:
 
 ## [Unreleased]
 * Added `.16byte` data directive to emit 16-byte values with configured multi-word endianness.
+* Added `#error` preprocessor directive for compile-time errors with optional message.
 
 ## [0.7.1]
 * Added support to the VSCode and Sublime Text language extensions for editor hovers for mnemonics, labels, and constants; plus improved semantic syntax highlighting that differentiations label and constant defintion from label and constant usage.